requirements.txt
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.4
hyperframe==5.2.0
idna==2.10
numpy==1.26.4
packaging==23.2
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.2
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
